refactor(dashboard): isolate exact-review queue service into its own module
